requirements.txt
ibott-robot-manager
ibott-browser-manager
ibott-files
pandas
openpyxl
PyAutoGUI
opencv_python
pywinauto
ibott-mail==1.0.6